Make SoapClient logging more flexible
This commit moves the log object in client.py into the SoapClient object
and adds a "log_id" parameter to its __init__ method.

By default it will use __name__ as its logger id, which will have the
same behavior it always has. Nothing changes.

However, you can either set log_id to your own custom logger id OR
set it to False, which will use a default id unique to that instance
of SoapClient (sets it to 'pysimplesoap_{0}'.format(id(self)) ).

With this change, calling SoapClient with "trace=True, log_id=False"
is a quick way to get soap traces for *only* that SoapClient instance.
This is useful if you are dealing with multiple servers, and only wish
to debug one of them.

Previously, soap tracing has been a module-level toggle due to the way
logger was configured.